added a geojson parser and import methods.
This new function will read all files which are thrown into the
data/external/geojson/import directory. All records in geojson files are
parsed and translate it into our internal message format. All files in
that directory must end with '.json'. Files which had been processed are
moved to data/external/geojson/imported. No of the imported records are
added to the message dump. If a loklak instance is reconstructed from
dumps, the geolocation dumps must be imported as well.

At this time, the import mechanism reads records from
http://www.paris-streetart.com/test/map.geojson
only. Because the 'property' object inside geojson is not defined
formally, the import process can only understand whats inside the
example file. This must be changed in the future.

Right now, there is one important attribute for the visualization of
markers from messages with geolocation: location_mark

There are I two locations in the json attached:
- the center of the location in location_point
- if the location is a point, the value in location_radiusis 0. If it is
large, there is a number which means meter. An example would be the
marker for a city with the radius of the city in this field
- a single location within the radius is in location_mark. This is set
randomly within the the radius of the location and shall be used for
visualization of the marker. The reason for using such a random field
is, that a location within a tweet can be very weak but we want to
calculate a location anyway. If we then have thousands of location for
i.e. Singapore, we don't want to have them overlapping in a single
point, but evenly distributed in a cloud around the center. To prevent
that these locations 'jump' every time you re-paint the map, they will
be computet once and you get that location in this field.
- the source of the location in location_source. This value will either
be REPORT if it was received from an external source or ANNOTATION if we
computed this location ourself using annotation techniques (i.e.
geo-database matching)