susi has moved out
it is now at https://github.com/fossasia/susi_server
Some remainings of susi must stay, because it had shared data structures
with the console service which stays in loklak. The console will be in
susi and loklak, but with different services. Susi will cross-reference
services from api.loklak.org as generic console service. So
loklakbecomes a microservice for susi.