Fix Query for POST requests
Instead of replacing qm when it's a POST request, it's better to
add the results rather than replacing all of the contents.

qm now contains the content of the request parameters and to reduce
code duplication, it's placed on the constructor.

initGET has been modified to complement the modification of the
constructor.

initPOST is deprecated because it's purpose has been replaced by the
constructor.

Fixes https://github.com/loklak/loklak_server/issues/913